requirements.txt
slack_sdk==3.26.0
python_dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
loguru==0.7.2